terraform: disable cloudfront root object for crates.io on staging
There is a concern that setting the default root object to index.html
for the crates.io webapp causes Fastboot to break:

    https://github.com/rust-lang/crates.io/pull/1937#issuecomment-570452017

This adds a temporary feature flag to remove the default root object
just from staging, without affecting production. After testing is done
the feature flag will have to be removed, either disabling or enabling
the root object both on staging and production.