Upgrade to terraform 1.0
This drops support for auto-opening ssh port forwarding for postgres
(rds-databases); that support seems to have stopped working and debugging it is
pretty hard. It seems like setting up the ssh channel on the side is not that
hard. It seems like broadly the consensus across various issue threads is that
this is hard or impossible to get right, and the answer may lie closer to either
a VPN like solution (probably not worth it for us) or moving terraform ops onto
the bastion itself.

The latter sounds like a potentially good idea; it gives us a concrete bouncing
off point and stable IP for various firewalls and place to provision the
relevant software to run terraform.

This commit also bumps the aws provider to 3.x in a couple places; this is
likely not strictly necessary but made it a little easier to integrate changes.
Ultimately seems like a good thing to keep up to date.