Bump rustc-perf to have some extra memory
We are seeing kills due to OOM on some of the self profile processing, which
currently requires that we hold a couple profiles in memory, and each can be
somewhat large. Ideally, we'd rearchitect to avoid that, but that is rather
difficult, and may not be generally feasible. For now just don't try that.
Some of the recent work to make perf's reports more reliable is also more
compute intensive; this should also help speed that up.