Single-machine crater configuration in terraform
This adds a GCP configuration to terraform which describes the initial workings
of our Crater agents there. Currently, this isn't an import of the existing
primary resources (which are still managed ad-hoc via the console), but it
represents the first step toward becoming that initial resource.

In a future commit, we'll rework the instance creation to use a Packer-created
GCP-AMI equivalent, rather than a startup script; the current configuration
hard-codes a particular instance (gcp-2) into the configuration, though still
fetching all credentials dynamically.

This is also an experiment with whether it's viable to run Crater on purely
ephemeral instances, where even the base disk is expected to vanish
periodically. Our hope is that the answer is "yes", as it'll simplify scaling
Crater up and down.

We'll also be adjusting the Crater code to support this kind of ephemeral
instance by moving away from specific tokens per-instance, which are difficult
to manage when instances are ephemeral. It's not quite clear what the right
answer is (e.g., we likely still want per-instance metrics), so it's possible
that we'll adjust to make the assignment there more dynamic but still leave
instance names.