Autoscale crater instances
Crater no longer requires each instance to be uniquely identified for
assigned crates to not be duplicated across instances, which means that we can
relatively directly just copy the stateless configuration into more instances.
This commit adds a CPU-based autoscaler to the GCP managed instance group, and
configures a systemd timer via the startup script which will update the
crater-agent on the instances.

This continues building upon the startup scripts rather than pursuing Ansible
automation (perhaps backed by Packer or so) because they're both pretty simple
and it doesn't seem worth the fairly high complexity cost Ansible and Packer
would bring at this time.

The current autoscaling is just so that in theory we shutdown instances during
idle periods (if we have them), but in practice that's likely more of an
opportunistic nice to have -- eventually it might allow us to burst up to finish
work much faster and then go idle, but in practice it's not clear whether we
want slow, but constant, or fast, but sometimes unavailable until the next
month. Something to continue thinking about.