Refactor to use managed cache policy
This avoids a super-long TTL on the woff2 files for now (setting to 24 hours),
which matches the TTL used by CloudFront today. In general this should probably
be enough for most purposes. We also add a stale-while-revalidate to let most
users check the resource in the background, rather than waiting on the request
to complete before proceeding.