Allow users to debug their processes
Ubuntu restricts ptrace by default, and only allows a process to debug
its own child processes. This is limiting the ability of developers to
use the dev desktops to analyze and debug their work.

The default scope for ptrace has been changed to `0`, allowing users to
debug their own processes. This carries with it the risk that malicious
code can interact with any process that a user created. We deem this to
be acceptable in this specific use case, as the dev desktops are
intended to be used only for the development of Rust and have, by
design, no user data or credentials that could be extracted.