Add (disabled) static.docs.rs distribution
This is intended to serve as an access point for a subset of S3 files,
currently configured via object tagging on the S3 side. We've disabled
it for now until we figure out whether we'll be using object tagging or
something else -- for example, cloudfront signed URLs.

Currently most of the objects in the bucket have a private ACL which
will prevent CloudFront access anyway; we expect that we'll drop the ACL
support on the bucket soon.