Install kernel-specific packages
The package required to enable user quotas installs kernel modules,
which will use the latest version of the kernel by default. This can
cause issues, since the system might be running an older version. In
this scenario, modprobe will fail to find the modules and return an
error.

Instead of only installing the latest version of the package, we are
now discovering the current version of the kernel and installing the
matching package as well. This ensures that quotas and perf work on the
current and an all future versions.