Rewrite scripts to set up Rust on dev desktops
The scripts to set up Rust for a user on the dev desktops have been
refactored for a better onboarding experience and error handling. The
init script installs rustup now, and can be extended in the future to
also set the user's git username and email.

For users who want to set up their own Rust toolchain, the setup_rust
script clones their fork of the rust-lang/rust repository, compiles it,
and links the build artifacts. This is optional, however, and not
required if users want to use the dev desktop to work with different
repositories.