Create script to migrate state
A Python script has been added that can be used to migrate state from
Terraform to terragrunt. It takes the name of a Terraform module as an
argument, and tries to import this module into the terragrunt
configuration in the current directory. Resources are limited by a
mandatory prefix, e.g. `module.staging`, to split and rename the
resources.

The script uses `terragrunt import` to migrate the resources. This means
no existing infrastructure or state is changed, but instead a duplicate
state is built up in terragrunt. Some resources and attributes cannot
get imported, though, and a `terragrunt apply` might still be required
at the end.