Increase quota on dev-desktops
A few users have reported running out of disk space on the new dev
desktops, which has a default file system quota of 50GB per user.
Increasing the quota to 100GB has resolved the issue for most users, so
we want to make this the new default to reduce friction for users and
work for the infra team.

With the increase in quota, two scripts have been added that allow
administrators to report the current quotas (`report-quotas.sh`) and to
increase them (`increase-quotas.sh`). These can be used on existing
machines to adjust the quota for all users to a new value.