Build Fastly function during terraform plan
The Fastly function is now built whenever `terraform plan` is run. This
This ensures that the function is always up to date, and prevents users
from accidentally uploading a stale version of the WASM module.

This is implemented using an external data source [^1] inside Terraform.
The data source calls a script that compiles the function and returns
the path to the artifact.

[^1]: https://registry.terraform.io/providers/hashicorp/external/latest/docs/data-sources/data_source