Make acm-certificate a sharable Terraform module
The acm-certificate module was falsely created as a service in
Terragrunt. But it is better suited to be a Terraform module that other
modules can include.

Due to slight differences between the legacy account and new accounts, a
new variable is necessary that changes how the DNS zone for the
certificate is calculated.