Decommission the original dev-desktop
The original dev-desktop that was used to develop the system
configuration has been deprecated. While we leave the machine itself
around for now to give users time to backup their data, we are already
removing it from Ansible and the monitoring. Instead, the four new
machines are now officially "live".