Add alternate name to Fastly service
The Fastly service for `static.crates.io` has its own domain and
certificate. This makes it easier to deploy and test the service, and it
allows us to split traffic using weighted DNS records in the future. For
this to work, though, the service needs to have a `static.crates.io` as
an alternate name in its certificate and it needs to respond to requests
for this domain.