Enable logging for static crates.io distribution
Request logs for the static crates distribution are now logged to an S3
bucket. Logs are either written from CloudFront or streamed from Fastly.

The setup on Fastly splits logs into request logs and service logs.
Request logs contain a single line for each request that is received,
while service logs provide internal logs of the cloud worker that are
useful when debugging issues.

Fastly is granted write-only access to the S3 bucket using a role. The
configuration is based on Fastly's documentation[^1].

[^1]: https://docs.fastly.com/en/guides/log-streaming-amazon-s3