Redirect database dump to CloudFront
An issue with the database dump for crates was reported on Zulip[^1],
and determined to be caused by the file size limit of Fastly. We can
currently cache files up to 50MB on Fastly, but the database dump is
close to 250MB in size. To avoid this issue, we are redirecting download
requests for the dump to CloudFront.

[^1]: https://rust-lang.zulipchat.com/#narrow/stream/242791-t-infra/topic/403s.20from.20.2Fdb-dump.2Etar.2Egz