Explicitly provide S3 endpoint
Fastly defaults to the S3 endpoint in us-east-1, but our buckets are in
us-west-1. Therefore, we have to set the S3 domain explicitly. We are
manually setting it to `s3.us-west-1.amazonaws.com`, because using the
bucket domain that Terraform returns would cause the paths inside the
bucket to be weird (i.e. `s3.amazonaws.com/bucket/bucket/path/to/log`).