Explicitly set CORS headers on Fastly
Two separate issues have been reported recently related to CORS headers
on Fastly. While we failed to reproduce the first[^1], the second[^2]
clearly points to different behavior between CloudFront and Fastly.
Namely, CloudFront always adds the CORS headers when the request
includes the `Origin` header. Fastly on the other hand seems to add the
headers only when hitting its internal domain `fastly-static.crates.io`
directly.

While we are investigating the issue with Fastly, we are setting the
CORS headers explicitly on requests that include the `Origin` header to
match the behavior of CloudFront.

[^1]: https://github.com/rust-lang/crates.io/issues/6080
[^2]: https://github.com/rust-lang/crates.io/issues/6164