Provision Global Sign certificate for Fastly
OpenSSL < 1.1 (and LibreSSL < 3.2) are not supported by Let's Encrypt,
which causes cargo to fail on older operating systems. To fix this, we
are switching crates over to Global Sign certificates. As a first step
towards that goal, a new certificate is being provisioned as part of the
crates.io module.