Bump promote-release job to 2xlarge (8 -> 72 cores)
We hit the previously configured 2 hour timeout with 8 cores while
recompressing. The quota for timeouts is 8 hours, and it's likely we'd
fit into that, but it's we don't really want to wait that long for
builds to complete (especially stable builds).

We may pursue alternatives (e.g., recompressing all files via highly
parallel Lambdas or something) but this seems like an OK quick fix.