Lock production environment to Git commit
The production environment for releases has been locked to a specific
Git commit. This makes it possible to test changes to the Terragrunt
modules in isolation on staging, and then promote them to production
once they have stabilized.

For consistency, the path for both the development and production
environment has been changed. This ensures that running `terragrunt` in
both environments produces the same directory structure in the cache.
This is extremely useful for debugging.