Update environment variable to support surrogate keys
The cache invalidation on Fastly for releases has been refactored to use
surrogate keys instead of paths[^1]. This removes the need to set the
static domain for Fastly, but requires passing in the respective service
id. The service ids for dev-static and static have already been
configured as SSM parameters.

[^1]: https://github.com/rust-lang/promote-release/pull/72