Deploy Wiz connector into AWS accounts
The integration with Wiz has been deployed into our AWS accounts. This
gives Wiz read-only access to our cloud infrastructure so that it can
perform vulnerability scanning.

When adding new accounts in the future, we need to deploy the Wiz
connector there as well. The alternative would have been to deploy Wiz
using a CloudFormation StackSet, but since we are not using
CloudFormation for anything else the decision was made to keep the
configuration in Terraform.