Run Podman in rootless mode on dev-desktops
Podman is a container runtime that supports running containers without
root access. This allows us to run it on a shared host like the
dev-desktops without compromising the security of the system.

The package for Podman and an alias for `docker` get installed through
the system's package manager, as are packages for user-space networking
and an alternative storage driver. The storage driver is enabled in a
global configuration file so that it applies to all users.