Disable monitoring of Lambda functions
Datadog can monitor Lambda functions, but charges a fee for each
function. This is quite expensive for us, since every routing function
we use in CloudFront is globally distributed and is billed multiple
times (once per region).

For now, we are just going to disable the Lambda integration. Later, we
might want to enable it again, but add a tag and filter to selectively
monitor specific functions.