Increase block size for segmented caching
Increasing the block size for segmented caching results in much better
performance, but might cause segments to be evicted from the cache
earlier and thus lowering the cache hit rate. Fastly does not recommend
raising the block size above 10MB for this exact reason. We are starting
at this limit to optimize for performance, and might lower the size if
we observe issues with the cache hit rate.

Read more: https://developer.fastly.com/reference/vcl/variables/segmented-caching/segmented-caching-block-size/