Use roles for RBAC
Instead of relying on teams to tag incoming resources and to use them
in restriction queries, we are using teams and the `service` or `app`
tag to control access to resources. The teams have been removed again,
and a more limited role for contributors and a "marker role" for the
crates.io team have been created.