Pass Accept-Encoding header to origin
The Accept-Encoding header is now passed to the origin so that it can
respond with compressed responses.

This fixes an inconsistency with the backend, which prefers Brotli
compression while CloudFront prefers gzip. Because the header is now
passed to the origin, the Brotli-compressed response from the origin is
passed through to the client. Previously, CloudFront would fetch
uncompressed resources and compress them on the fly using gzip.