Remove unused cache behavior
The custom cache behavior for `/assets` was never used, since the
pattern matches only the specific `/assets` URL and not any resources
served under it (e.g. `/assets/foo`). This was discovered while
investigating the inconsistencies in #356.

Since the configuration was never used in the first place, we are
removing it to remove some complexity.