Remove billing access from legacy account
The staff of the Rust Foundation has access to the billing portal on
AWS, historically through the legacy account and now through the new
root account that consolidates billing. Instead of updating the custom
policy to remove deprecated policies, we are going to remove billing
access from the legacy altogether since it is no longer required.

Fixes #359