Deploy SQS queue for crates.io to staging
The SQS queue that was configured in #377 has been deployed to the new
staging account for crates.io that was created in #374. Slight
modifications were necessary to the configuration:

  - The resource and human-readable names of the SQS are now identical.
  - The `sid` for policies matches the naming rules of AWS.
  - The input variable has been changed, since the account number is not
    part of a bucket's ARN and can thus not be extracted from it.

The infrastructure has been deployed with the same version of the
Terraform provider for AWS as the other modules in simpleinfra to ensure
future compatibility.