Move resources into the right region
The SQS queue must be in the same region as the S3 bucket, otherwise it
cannot be configured as the target for notifications. With our current
Terragrunt setup, it is not possible to create resources in a different
region, so we had to globally change the region that is associated with
this account.