Clean up unused checkouts on dev-desktops
User might checkout repositories on the dev-desktops, work on them for a
while, and then move on to other projects. The unused checkouts, in
particular their `build` or `target` directories, can use up a lot of
space that we want to reclaim periodically. A script has been created
that finds all cache directories, checks recursively if any file in
their root directory has been modified within a given time frame, and if
not deletes the cache directory.