Enable metadata for Fastly function again
The collection of metadata was disabled in #387, because the metadata
contained an unstable property that broke the change detection in
Terraform. This issue has been resolved upstream by introducing buckets
for the heap allocation size, which reduces the risk of changes in the
metadata while preserving the usefulness of the property.

See https://github.com/fastly/cli/pull/1130 for details on the buckets.