Move hard-coded constants into Compute@Edge module
In an effort to simplify the configuration of both the Terraform and
Rust modules, some hard-coded constants have been moved into the
Compute@Edge module. This makes it possible to remove the glue code to
pass them from the Terraform configuration to the final WASM function.
While this does introduce some duplication, it removes a lot of
complexity and potential configuration issues.