Create bucket to store downloads statistics
The crates.io team is considering adding a long-term archive for
download statistics so that historic data can be removed from the
database. The proposal in rust-lang/crates.io#3479 is to put daily CSV
files into an S3 bucket.

A new bucket has been created for this purpose. Other than the existing
buckets, this one is not publicly accessible. The bucket has been
created in the new `crates-io-staging` account so that the team can more
easily access it.