Support multiple AWS regions in Terragrunt
Some of our Terraform modules deploy resources to different regions
simultaneously. This is done by defining multiple `provider` blocks for
AWS with different regions and different aliases, which can then be
references in the Terraform resources.

With Terragrunt, we are dynamically creating the `provider` block and
overwriting any additional providers that are defined in Terraform
files. Our script has therefore been extended to accept multiple regions
in the `account.json` file and expand them to the same structure that we
were using before with Terraform.