Tag rustup artifacts with surrogate key
We use surrogate keys[^1] on Fastly to invalidate the cache for specific
objects, e.g. when new versions of Rust or rustup are released.
Surrogate keys are transparently set in the Fastly configuration based
on the request path.

A new VCL snippet has been added that tags every rustup artifact with
the `rustup` surrogate key so that we can purge them from the cache when
a new rustup version is released.