Create Terragrunt module for win.rustup.rs
`win.rustup.rs` is an old CloudFront distribution that is not yet
managed in Terraform or Terragrunt. A new Terragrunt module has been
created so that we can easily create a new staging environment.

The module copies the existing functionality, but uses a response header
policy instead of a Lambda function to set the `Content-Disposition`
header.