Allow the crates.io team to manage log pipelines
We use Datadog among other reasons for the application logs from
crates.io. As new features or services are added, the crates.io team
reguarly wants to tweak the pipelines slightly to improve the parsing of
logs and extract data into metrics. This requires elevated permissions
to the log pipelines and processors. While ideally, we would only grant
these permissions for the specific pipeline used by crates.io, this is
not possible through the API as far as we know.