Expand Datadog permissions
More permissions on Datadog have been granted.

Notebooks can be a useful feature for teams to document their processes
or create runbooks for debugging. They are free to use, so the
permission has been granted to all teams.

The staff of the Rust Foundation has additionally been granted the
permissions to create metrics from logs, edit their metadata, and share
dashboards publicly. These permissions are limited to staff for now,
because custom metrics are a paid feature and we want to limit who can
share data publicly.