Create S3 bucket for Rustup build artifacts
A new S3 bucket is being created as part of the effort to rebuild the
Rustup release process. The bucket will store the build artifacts for
Rustup, which the GitHub Actions in rust-lang/rustup will produce.

Since the bucket is tied to a GitHub repository, only a single bucket in
the production environment is being created.