Remove warning for unnecessary lifecycle rule
The `latest_restorable_time` cannot be compared between Terraform plans
anyways and thus doesn't need to be ignored. Terraform started printing
a warning for this configuration, so the attribute has been removed. In
its place, the engine version of the database is being ignored since it
is periodically updated manually through the AWS console.