Remove IP filter from bastion's security group
In the past, access to the bastion host was restricted to static IP
addresses belonging to the users with access. This worked well for users
with a (semi-)static IP address, but introduced a lot of friction for
users on most residential internet connections in Europe (e.g. when
accessing the server from their home). Due to this friction, we have
decided[^1] to drop the IP filter.

[^1]: https://rust-lang.zulipchat.com/#narrow/stream/242791-t-infra/topic/drop.20the.20bastion.20IP.20requirement.2C.20or.20idea.20for.20a.20better.20way.2E.2E.2E