Switch to smaller instance type
We ran into long QUEUED times with the larger instances, and in testing
it seems like the smaller instance type isn't actually significantly
slower, and is 2x less expensive, so we should switch to it anyway.

This also bumps the build timeout to give us a bit more buffer (our
builds are currently just under 2 hours).