Prevent infinite redirect for doc-router /doc[^/]
Previously, it would just put the slice in the Location with no regards
for what was in there. If there was no slash after the `/doc`, it would
issue a relative redirect, which was still under `/doc`, causing an
infinite redirect.

- /docs/std -> s/std
- /docs/s/std -> s/s/std
- /docs/s/s/s/std -> ...

We prevent this behavior here by only applying the `/doc` redirect if
there is a slash directly after it, causing the default /stable rewrite
fallback to be applied to everything else, which will in practice lead
to nice 404 page.

This should not break anything as all such queries currently run into
these loops, including plain `/doc`.